refactored utility functions into utils.py and added some documentation
