+++ .gitignore
@@ -5,3 +5,5 @@
 /build/
 /dist/
 /mkalias.egg-info/
+*.DS_Store
+/mkalias_foss.egg-info/
